recreate pipelines on certain errors (#5)
This will recreate pipelines when the spinnaker API returns the
following error on pipeline creation or update operations:

    A pipeline with name <pipeline-name> already exists

We handle this by just recreating the problematic pipelines because
there is no other feasible way how to resolve this issue. The issue is
especially annoying because we also sometime see this error on update
operations where it is expected that the pipeline exists.