fix: lazily initialize gate API client (#10)
The initialization of the gate API client results in an immediate API
call to the `/version` endpoint. This makes it impossible to use the
provider inside of a module that will create the spinnaker as well
because the presence of the `provider "spinnaker" {}` block alone will
result in an API call to a non-existent endpoint.

This change delays the initialization of the gate API client until it is
needed. That allows the creation of the spinnaker and resources managed
by the spinnaker provider inside the same module by leveraging
`depends_on`.

The change also adds a `Makefile` to build and install the provider for
local testing as this has become a pain starting with terraform 0.13.