chore: cleanup codebase (#11)
This change cleans up the code base in order to simplify adding new
features in the future. Does not introduce any behaviour changes or
provider features yet.

Since the original provider (https://github.com/armory-io/terraform-provider-spinnaker)
is no longer actively maintained this change removes the bits that where
specific to the origin repository.

Notable changes:

- Package import path changed to `github.com/Bonial-International-GmbH/terraform-provider-spinnaker`
- Armory CircleCI config removed. Instead a github actions build
  workflow was added.
- CircleCI badge in `README.md` replaced with project specific ones
- Bonial International GmbH was added to the `LICENSE`
- `.golangci.yml` config added
- `go fmt` was run across all source files
- Go version was bumped to 1.16 (latest)
- Linting issues and broken tests were fixed

After this change is merged the repo should be in a good enough state to
add more features and test, and to update all current dependencies to
the latest versions (possibly via dependabot).